refactor(storage): port root writes to shared serializer
